Q&A PDF ChatBot/requirements.txt
langchain==0.0.27
PyPDF2==3.0.1
Python-dotenv==0.20.0
streamlit==1.10.0
faiss-cpu==1.7.4